# gpfdist: set default compression level to 3