## 🔍 Test Configuration
- **Test tool**: wrk
- **Test parameters**: 4 threads, 10 connections, 15s duration

## 📊 Complete Performance Comparison Table

| Configuration | RPS | Relative Baseline Gain | Avg Latency | P50 Latency | P90 Latency | P99 Latency | Transfer Rate |
|---------------|-----|-----------------------|-------------|-------------|-------------|-------------|---------------|
| No Compression (Baseline) | 12,151.35 | 0.00% | 572.99μs | 563.00μs | 657.00μs | 748.00μs | 2.19MB |
| Compression Level 1 | 12,169.58 | +0.15% | 583.22μs | 559.00μs | 654.00μs | 773.00μs | 2.20MB |
| Compression Level 2 | 12,237.98 | +0.71% | 565.10μs | 562.00μs | 652.00μs | 719.00μs | 2.21MB |
| **Compression Level 3** | **12,254.67** | **+0.85%** | **571.59μs** | **559.00μs** | **652.00μs** | **725.00μs** | **2.21MB** |
| Compression Level 4 | 12,248.22 | +0.80% | 573.53μs | 556.00μs | 651.00μs | 748.00μs | 2.21MB |
| Compression Level 5 | 12,067.48 | -0.69% | 580.70μs | 566.00μs | 661.00μs | 754.00μs | 2.18MB |
| Compression Level 6 | 12,185.62 | +0.28% | 575.21μs | 558.00μs | 653.00μs | 753.00μs | 2.20MB |
| Compression Level 7 | 12,182.25 | +0.25% | 574.92μs | 563.00μs | 654.00μs | 734.00μs | 2.20MB |
| Compression Level 8 | 12,177.17 | +0.21% | 575.67μs | 566.00μs | 652.00μs | 718.00μs | 2.20MB |
| Compression Level 9 | 11,741.52 | -3.37% | 595.62μs | 586.00μs | 684.00μs | 815.00μs | 2.12MB |

## 🔎 Key Findings
1. **Optimal Level**: Level 3 shows best overall performance (+0.85% RPS)
2. **Diminishing Returns**: Higher levels (7-9) show reduced benefits
3. **Performance Drop**: Level 9 shows significant degradation (-3.37% RPS)
4. **Latency Impact**: Compression generally maintains or improves latency until level 5+


Fixes #ISSUE_Number

### What does this PR do?
<!-- Brief overview of the changes, including any major features or fixes -->

### Type of Change
- [ ] Bug fix (non-breaking change)
- [ ] New feature (non-breaking change)
- [ ] Breaking change (fix or feature with breaking changes)
- [ ] Documentation update

### Breaking Changes
<!-- Remove if not applicable. If yes, explain impact and migration path -->

### Test Plan
<!-- How did you test these changes? -->
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated
- [ ] Passed `make installcheck`
- [ ] Passed `make -C src/test installcheck-cbdb-parallel`

### Impact
<!-- Remove sections that don't apply -->
**Performance:**
<!-- Any performance implications? -->

**User-facing changes:**
<!-- Any changes visible to users? -->

**Dependencies:**
<!-- New dependencies or version changes? -->

### Checklist
- [ ] Followed [contribution guide](https://cloudberry.apache.org/contribute/code)
- [ ] Added/updated documentation
- [ ] Reviewed code for security implications
- [ ] Requested review from [cloudberry committers](https://github.com/orgs/apache/teams/cloudberry-committers)

### Additional Context
<!-- Any other information that would help reviewers? Remove if none -->

### CI Skip Instructions
<!--
To skip CI builds, add the appropriate CI skip identifier to your PR title.
The identifier must:
- Be in square brackets []
- Include the word "ci" and either "skip" or "no"
- Only use for documentation-only changes or when absolutely necessary
-->

---
<!-- Join our community:
- Mailing list: [dev@cloudberry.apache.org](https://lists.apache.org/list.html?dev@cloudberry.apache.org) (subscribe: dev-subscribe@cloudberry.apache.org)
- Discussions: https://github.com/apache/cloudberry/discussions -->